slight mergin and refactor; fixed tsreconstruct driver
